requirements.txt
# Eden Analytics Pro - Requirements
# Version 3.0.0 Monster Edition - ML Model v5.0

# ========================================
# Core Dependencies
# ========================================
requests>=2.28.0
pyyaml>=6.0
pydantic>=2.0.0

# ========================================
# GUI Framework
# ========================================
PyQt6>=6.4.0
PyQt6-WebEngine>=6.4.0

# Video playback for splash screen
opencv-python>=4.8.0

# ========================================
# Web Scraping (Belarusian bookmakers)
# ========================================
beautifulsoup4>=4.12.0
lxml>=4.9.0

# ========================================
# Data Science & ML
# ========================================
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
matplotlib>=3.6.0

# ========================================
# Deep Learning
# ========================================
tensorflow>=2.13.0
keras>=3.0.0

# ========================================
# Gradient Boosting Ensemble
# ========================================
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0

# ========================================
# Hyperparameter Optimization
# ========================================
optuna>=3.3.0

# ========================================
# Model Interpretability
# ========================================
shap>=0.42.0

# ========================================
# Interactive Charts
# ========================================
plotly>=5.15.0

# ========================================
# Scheduling
# ========================================
apscheduler>=3.10.0

# ========================================
# Telegram Bot
# ========================================
python-telegram-bot>=20.0

# ========================================
# Export Formats
# ========================================
openpyxl>=3.1.0
reportlab>=4.0.0

# ========================================
# Console UI
# ========================================
rich>=13.0.0
